Shader Color
The Shader Color is updating with the change of the categories tabs.

Model component is now in a separate file including the shader ft the animation

Restructured some the of the functions so I can easily prop it down the component tree

Next:
-Page Transitions
-Change in the model itself
-Model Change Transition
