+++ package.json
@@ -18,6 +18,7 @@
     "react-locomotive-scroll": "^0.2.2",
     "react-responsive-carousel": "^3.2.23",
     "react-slick": "^0.29.0",
+    "react-spring": "^9.7.3",
     "react-use-clock": "^0.0.4",
     "sass": "^1.66.1",
     "scss": "^0.2.4",
